Fix: 불필요한 validation 메시지 제거
