--- src/utils/validation.ts
@@ -1,29 +1,0 @@
-import { DocumentData, QueryDocumentSnapshot } from 'firebase/firestore';
-import { z } from 'zod';
-
-export type ErrorCode = {
-  [key: string]: string;
-};
-
-// 에러 메세지
-export const ERROR_MESSAGES: ErrorCode = {
-  INVALID_ACCOUNT: '이메일과 비밀번호를 확인해주세요.',
-};
-
-// 로그인 유효성 검사
-export function validateLogIn(
-  docs: QueryDocumentSnapshot<DocumentData, DocumentData>[]
-) {
-  if (docs.length === 0) return ERROR_MESSAGES.INVALID_ACCOUNT;
-
-  return null;
-}
-
-const Login = z.object({
-  name: z.string(),
-  height: z.number(),
-  age: z.number(),
-  phoneNum: z.string(),
-  homePhoneNum: z.string().optional(),
-  isCompletedMilitaryService: z.boolean(),
-});
